requirements.txt
astroid==2.5.2
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cmake==3.18.4.post1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
distlib==0.3.3
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Faker==9.8.2
filelock==3.3.1
Flask==2.0.2
geographiclib==1.52
geopy==2.2.0
google-api-core==1.26.3
google-auth==1.28.0
google-cloud==0.34.0
google-cloud-vision==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
imageio==2.10.3
imutils==0.5.4
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
jsonify==0.5
jupyter-client==6.1.12
jupyter-core==4.7.1
kaggle==1.5.12
keyboard==0.13.5
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
networkx==2.6.3
numpy==1.20.2
opencv-contrib-python==4.5.4.58
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
playsound==1.3.0
prompt-toolkit==3.0.19
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pydub==0.25.1
pygame==2.0.2
PyGetWindow==0.0.9
Pygments==2.9.0
PyJWT==2.3.0
pylint==2.7.4
PyMsgBox==1.0.9
pynput==1.7.3
pyodbc==4.0.32
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pyshine==0.0.9
python-dateutil==2.8.1
python-dotenv==0.19.2
python-slugify==5.0.2
python-vlc==3.0.12118
PyTweening==1.0.3
pytz==2021.1
PyWavelets==1.2.0
pywin32==301
pyzmq==22.1.0
requests==2.25.1
robotframework==4.1
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
sklearn==0.0
text-unidecode==1.3
threadpoolctl==2.2.0
tifffile==2021.11.2
tk==0.1.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
twilio==7.3.1
urllib3==1.26.4
virtualenv==20.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
wxPython==4.1.1
